ipatests: create group with gid inside the id range
The auto-private-group tests check the behavior of SSSD
when auto-private-group is configured AND an idoverrideuser
is set.
The test creates a group and uses the group gid for the
idoverrideuser, but the gid must be inside a local range
(otherwise the group doesn't have any SID).
The test used to force a gid value which could be outside
of the local IPA range and this resulted in unexpected
results on the client.
Instead of forcing a gid value during group creation, let
IPA assign a gid inside the existing range.

Related: https://pagure.io/freeipa/issue/9295

Signed-off-by: Florence Blanc-Renaud <flo@redhat.com>