Installation: use absolute DNS hostname to find IP addresses
During ipa server installation, the IP addresses are found by
calling get_server_ip_address(host_name). This method internally
calls resolve_ip_addresses_nss but returns only IPv4 addresses
if the host name is not an absolute DNS name (ending with a dot).
This list of IP addresses is later used to create the DNS records
for the server (and is missing the AAAA records).

Later on, the DNS server is setup and the records for ipa-ca are
added, using the IP addresses returned by resolve_ip_addresses_nss,
but this time called with an absolute DNS name, thus containing
both IPv4 and IPv6 addresses.

This results in an inconsistent list of DNS records as the server
has only IPv4 addresses but ipa-ca has both IPv4 and IPv6 addresses.
The problem is reported by ipa-healthcheck.

Fix the problem by using an absolute DNS name in the first step, so
that both IPv4 and IPv6 addresses are found and A and AAAA records
are created for the server.