Retrieve and add all CA certificates when verifying custom
In ipa-server-certinstall we validate the certificate chain
in a temporary NSS db after adding the new cert and the
CA chain. The issue was that if there were duplicate subjects
which causes certutil -L -a to return them as a series of PEM
files only the first one was being loaded for verification. This
could cause false-positive verification failures.

So for this case load all of them.

Note that besides this separate trust for different certs with
the same nickname are not allowed. Doubly so here because the
last one will win.

Related: https://pagure.io/freeipa/issue/9652

Signed-off-by: Rob Crittenden <rcritten@redhat.com>